Reorganize electron folder
