+++ gui/electron/main.js
@@ -2,7 +2,7 @@ const { app, BrowserWindow } = require('electron');
 const isDev = require('electron-is-dev');
 const path = require("path");
 
-const { buildMenu } = require('./menu.js');
+const { buildMenu } = require('./utils/menu.js');
 
 function createWindow() {
   // Create the browser window.
