+++ gui/electron/res/orbit-icon.png
