+++ gui/electron/utils/popOutData.js
